killing more runaway threads that are trying to work outside the test scope
