+++ kifi-backend/modules/common/app/com/keepit/common/mail/PlayMailerPlugin.scala
@@ -1,5 +1,7 @@
 package com.keepit.common.mail
 
+import net.codingwell.scalaguice.InjectorExtensions._
+
 import com.keepit.FortyTwoGlobal
 import com.typesafe.plugin.{ MailerAPI, MailerPlugin }
 
@@ -9,6 +11,6 @@ class PlayMailerPlugin(app: Application) extends MailerPlugin {
   def email: MailerAPI = {
     val global = app.global.asInstanceOf[FortyTwoGlobal]
     implicit val injector = global.injector
-    global.inject[MailerAPI]
+    global.injector.instance[MailerAPI]
   }
 }
