+++ kifi-backend/modules/shoebox/test/com/keepit/controllers/admin/AdminAuthControllerTest.scala
@@ -1,6 +1,6 @@
 package com.keepit.controllers.admin
 
-import com.keepit.common.concurrent.ExecutionContextModule
+import com.keepit.common.concurrent.{ FakeExecutionContextModule, ExecutionContextModule }
 import com.keepit.common.controller.{ FakeUserActionsHelper, FakeUserActionsModule }
 import com.keepit.curator.FakeCuratorServiceClientModule
 import org.specs2.mutable.Specification
@@ -30,7 +30,7 @@ import com.keepit.cortex.FakeCortexServiceClientModule
 class AdminAuthControllerTest extends Specification with ShoeboxApplicationInjector {
 
   val modules = Seq(
-    ExecutionContextModule(),
+    FakeExecutionContextModule(),
     FakeUserActionsModule(),
     FakeShoeboxServiceModule(),
     FakeScrapeSchedulerModule(),
