+++ kifi-backend/modules/shoebox/test/com/keepit/controllers/admin/AdminDashboardControllerTest.scala
@@ -1,6 +1,6 @@
 package com.keepit.controllers.admin
 
-import com.keepit.common.concurrent.ExecutionContextModule
+import com.keepit.common.concurrent.{ FakeExecutionContextModule, ExecutionContextModule }
 import com.keepit.curator.FakeCuratorServiceClientModule
 import org.joda.time.DateTime
 import org.specs2.mutable.Specification
@@ -31,7 +31,7 @@ import scala.concurrent.Future
 class AdminDashboardControllerTest extends Specification with ShoeboxApplicationInjector {
 
   def requiredModules = Seq(
-    ExecutionContextModule(),
+    FakeExecutionContextModule(),
     FakeUserActionsModule(),
     FakeSearchServiceClientModule(),
     FakeScrapeSchedulerModule(),
